refactor(#60/curation): 테스트 환경에서 S3 제외
